move `signal` into its own module
