+++ tests/parallel/main.rs
@@ -3,4 +3,5 @@ mod setup;
 mod cancellation;
 mod independent;
 mod race;
+mod signal;
 mod true_parallel;
